+++ api/tests/unit_tests/core/plugin/impl/test_model_runtime_factory.py
@@ -1,6 +1,7 @@
 from unittest.mock import Mock, patch
 
 from core.plugin.impl.model_runtime_factory import create_plugin_model_assembly
+from core.plugin.plugin_service import PluginService
 
 
 def test_plugin_model_assembly_reuses_single_runtime_across_views():
@@ -34,3 +35,11 @@ def test_plugin_model_assembly_reuses_single_runtime_across_views():
     mock_provider_factory_cls.assert_called_once_with(runtime=runtime)
     mock_provider_manager_cls.assert_called_once_with(model_runtime=runtime)
     mock_model_manager_cls.assert_called_once_with(provider_manager=provider_manager)
+
+
+def test_create_plugin_model_runtime_injects_plugin_service():
+    from core.plugin.impl.model_runtime_factory import create_plugin_model_runtime
+
+    runtime = create_plugin_model_runtime(tenant_id="tenant-1", user_id="user-1")
+
+    assert runtime._plugin_service is PluginService
